+++ scripts/generate_corpus.py
@@ -1,4 +1,4 @@
-from sqlschm.parser import parse_schema
+from sqlschm import parse_schema, generate_schema, Dialect
 import black
 from black import mode
 import os
@@ -12,13 +12,20 @@ def main():
     files.sort()
     for schm_name in files:
         if schm_name.endswith(".sql"):
-            print("Generating... " + schm_name)
-            outFile = Path(CORPUS + schm_name[:-4] + ".ast")
-            outFile.touch(exist_ok=True)
-            with open(CORPUS + schm_name) as schm, open(outFile, "w+") as out:
+            print("Generating... " + schm_name[:-4])
+            astFilename = Path(CORPUS + schm_name[:-4] + ".ast")
+            astFilename.touch(exist_ok=True)
+            outFilename = Path(CORPUS + schm_name[:-4] + ".out")
+            outFilename.touch(exist_ok=True)
+            with open(CORPUS + schm_name) as schm, open(
+                astFilename, "w+"
+            ) as astFile, open(CORPUS + schm_name) as schm, open(
+                outFilename, "w+"
+            ) as outFile:
                 content = schm.read()
-                ast = parse_schema(content)
-                out.write(black.format_str(repr(ast), mode=mode.Mode()))
+                schema = parse_schema(content)
+                astFile.write(black.format_str(repr(schema), mode=mode.Mode()))
+                outFile.write(generate_schema(schema, Dialect.SQLITE))
 
 
 if __name__ == "__main__":