+++ sqlschm/__init__.py
@@ -1,1 +1,5 @@
-__version__ = "0.0.0"
+# type: ignore
+from .generator import generate_schema
+from .parser import parse_schema
+from .sql import *
+from . import lexer, tok