+++ tests/test_sqlschm.py
@@ -1,5 +1,0 @@
-from sqlschm import __version__
-
-
-def test_version():
-    assert __version__ == '0.0.0'